+++ src/logic/enemies/enemy.cpp
@@ -2,7 +2,7 @@
 
 namespace td {
 Enemy::Enemy(types::Position position, float hitbox, types::Texture texture,
-             float health, int move_speed, float bounty, bool armored,
+             float health, int move_speed, int bounty, bool armored,
              float distance_moved)
     : Object(position, hitbox, texture),
       health_(health),
@@ -17,7 +17,7 @@ float Enemy::getHealth() const { return health_; }
 
 int Enemy::getMoveSpeed() const { return move_speed_; }
 
-float Enemy::getBounty() const { return bounty_; }
+int Enemy::getBounty() const { return bounty_; }
 
 bool Enemy::isArmored() const { return armored_; }
